Give a red hue to the detected region.
* src/manualRosetteDetect/findSegmentedRosette.m(findSegmentedRosette): Cast
the coordinates to double instead of int32. This allows for summation with
doubles in rosettedetect_gui.m.
* src/manualRosetteDetect/rosettedetect_gui.m(segment_Callback): Find the
corresponding coordinates in total image and force the red dimension to 255.

+++ src/manualRosetteDetect/findSegmentedRosette.m
@@ -15,7 +15,7 @@
 % Foundation, Inc., 675 Mass Ave, Cambridge, MA 02139, USA.
 function [subimg, imgOffset] = findSegmentedRosette ( lh, img )
     % rosette center
-    rc = uint32(round(lh));
+    rc = double(round(lh));
 
     % true when we are satified with the mask
     foundRosette = false;
